Move TxSet, TxSize and TxType to the transform module
