+++ src/context.rs
@@ -20,9 +20,10 @@ use crate::header::ReferenceMode;
 use crate::partition::BlockSize::*;
 use crate::partition::PredictionMode::*;
 use crate::partition::RefType::*;
-use crate::partition::TxSize::*;
-use crate::partition::TxType::*;
 use crate::partition::*;
+use crate::transform::TxSize::*;
+use crate::transform::TxType::*;
+use crate::transform::*;
 use crate::lrf::*;
 use crate::plane::*;
 use crate::scan_order::*;
