+++ src/partition.rs
@@ -19,6 +19,7 @@ use crate::mc::*;
 use crate::plane::*;
 use crate::predict::*;
 use crate::tiling::*;
+use crate::transform::TxSize;
 use crate::util::*;
 
 // LAST_FRAME through ALTREF_FRAME correspond to slots 0-6.
@@ -367,189 +368,6 @@ impl BlockSize {
   }
 }
 
-/// Transform Size
-#[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
-#[repr(C)]
-pub enum TxSize {
-  TX_4X4,
-  TX_8X8,
-  TX_16X16,
-  TX_32X32,
-  TX_64X64,
-
-  TX_4X8,
-  TX_8X4,
-  TX_8X16,
-  TX_16X8,
-  TX_16X32,
-  TX_32X16,
-  TX_32X64,
-  TX_64X32,
-
-  TX_4X16,
-  TX_16X4,
-  TX_8X32,
-  TX_32X8,
-  TX_16X64,
-  TX_64X16
-}
-
-impl TxSize {
-  /// Number of square transform sizes
-  pub const TX_SIZES: usize = 5;
-
-  /// Number of transform sizes (including non-square sizes)
-  pub const TX_SIZES_ALL: usize = 14 + 5;
-
-  pub fn width(self) -> usize {
-    1 << self.width_log2()
-  }
-
-  pub fn width_log2(self) -> usize {
-    match self {
-      TX_4X4 | TX_4X8 | TX_4X16 => 2,
-      TX_8X8 | TX_8X4 | TX_8X16 | TX_8X32 => 3,
-      TX_16X16 | TX_16X8 | TX_16X32 | TX_16X4 | TX_16X64 => 4,
-      TX_32X32 | TX_32X16 | TX_32X64 | TX_32X8 => 5,
-      TX_64X64 | TX_64X32 | TX_64X16 => 6
-    }
-  }
-
-  pub fn width_index(self) -> usize {
-    self.width_log2() - TX_4X4.width_log2()
-  }
-
-  pub fn height(self) -> usize {
-    1 << self.height_log2()
-  }
-
-  pub fn height_log2(self) -> usize {
-    match self {
-      TX_4X4 | TX_8X4 | TX_16X4 => 2,
-      TX_8X8 | TX_4X8 | TX_16X8 | TX_32X8 => 3,
-      TX_16X16 | TX_8X16 | TX_32X16 | TX_4X16 | TX_64X16 => 4,
-      TX_32X32 | TX_16X32 | TX_64X32 | TX_8X32 => 5,
-      TX_64X64 | TX_32X64 | TX_16X64 => 6
-    }
-  }
-
-  pub fn height_index(self) -> usize {
-    self.height_log2() - TX_4X4.height_log2()
-  }
-
-  pub fn width_mi(self) -> usize {
-    self.width() >> MI_SIZE_LOG2
-  }
-
-  pub fn area(self) -> usize {
-    1 << self.area_log2()
-  }
-
-  pub fn area_log2(self) -> usize {
-    self.width_log2() + self.height_log2()
-  }
-
-  pub fn height_mi(self) -> usize {
-    self.height() >> MI_SIZE_LOG2
-  }
-
-  pub fn block_size(self) -> BlockSize {
-    match self {
-      TX_4X4 => BLOCK_4X4,
-      TX_8X8 => BLOCK_8X8,
-      TX_16X16 => BLOCK_16X16,
-      TX_32X32 => BLOCK_32X32,
-      TX_64X64 => BLOCK_64X64,
-      TX_4X8 => BLOCK_4X8,
-      TX_8X4 => BLOCK_8X4,
-      TX_8X16 => BLOCK_8X16,
-      TX_16X8 => BLOCK_16X8,
-      TX_16X32 => BLOCK_16X32,
-      TX_32X16 => BLOCK_32X16,
-      TX_32X64 => BLOCK_32X64,
-      TX_64X32 => BLOCK_64X32,
-      TX_4X16 => BLOCK_4X16,
-      TX_16X4 => BLOCK_16X4,
-      TX_8X32 => BLOCK_8X32,
-      TX_32X8 => BLOCK_32X8,
-      TX_16X64 => BLOCK_16X64,
-      TX_64X16 => BLOCK_64X16
-    }
-  }
-
-  pub fn sqr(self) -> TxSize {
-    match self {
-      TX_4X4 | TX_4X8 | TX_8X4 | TX_4X16 | TX_16X4 => TX_4X4,
-      TX_8X8 | TX_8X16 | TX_16X8 | TX_8X32 | TX_32X8 => TX_8X8,
-      TX_16X16 | TX_16X32 | TX_32X16 | TX_16X64 | TX_64X16 => TX_16X16,
-      TX_32X32 | TX_32X64 | TX_64X32 => TX_32X32,
-      TX_64X64 => TX_64X64
-    }
-  }
-
-  pub fn sqr_up(self) -> TxSize {
-    match self {
-      TX_4X4 => TX_4X4,
-      TX_8X8 | TX_4X8 | TX_8X4 => TX_8X8,
-      TX_16X16 | TX_8X16 | TX_16X8 | TX_4X16 | TX_16X4 => TX_16X16,
-      TX_32X32 | TX_16X32 | TX_32X16 | TX_8X32 | TX_32X8 => TX_32X32,
-      TX_64X64 | TX_32X64 | TX_64X32 | TX_16X64 | TX_64X16 => TX_64X64
-    }
-  }
-
-  pub fn by_dims(w: usize, h: usize) -> TxSize {
-    match (w, h) {
-      (4, 4) => TX_4X4,
-      (8, 8) => TX_8X8,
-      (16, 16) => TX_16X16,
-      (32, 32) => TX_32X32,
-      (64, 64) => TX_64X64,
-      (4, 8) => TX_4X8,
-      (8, 4) => TX_8X4,
-      (8, 16) => TX_8X16,
-      (16, 8) => TX_16X8,
-      (16, 32) => TX_16X32,
-      (32, 16) => TX_32X16,
-      (32, 64) => TX_32X64,
-      (64, 32) => TX_64X32,
-      (4, 16) => TX_4X16,
-      (16, 4) => TX_16X4,
-      (8, 32) => TX_8X32,
-      (32, 8) => TX_32X8,
-      (16, 64) => TX_16X64,
-      (64, 16) => TX_64X16,
-      _ => unreachable!()
-    }
-  }
-
-  pub fn is_rect(self) -> bool {
-    self.width_log2() != self.height_log2()
-  }
-}
-
-pub const TX_TYPES: usize = 16;
-
-#[derive(Debug, Copy, Clone, PartialEq, PartialOrd)]
-#[repr(C)]
-pub enum TxType {
-  DCT_DCT = 0,   // DCT  in both horizontal and vertical
-  ADST_DCT = 1,  // ADST in vertical, DCT in horizontal
-  DCT_ADST = 2,  // DCT  in vertical, ADST in horizontal
-  ADST_ADST = 3, // ADST in both directions
-  FLIPADST_DCT = 4,
-  DCT_FLIPADST = 5,
-  FLIPADST_FLIPADST = 6,
-  ADST_FLIPADST = 7,
-  FLIPADST_ADST = 8,
-  IDTX = 9,
-  V_DCT = 10,
-  H_DCT = 11,
-  V_ADST = 12,
-  H_ADST = 13,
-  V_FLIPADST = 14,
-  H_FLIPADST = 15
-}
-
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub enum PredictionMode {
   DC_PRED,     // Average of above and left pixels
@@ -591,6 +409,7 @@ pub enum InterIntraMode {
   II_SMOOTH_PRED,
   INTERINTRA_MODES
 }
+
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub enum CompoundType {
   COMPOUND_AVERAGE,
@@ -693,16 +512,6 @@ pub const REFMV_CTX_MASK: usize = ((1 << (8 - REFMV_OFFSET)) - 1);
 pub static RAV1E_PARTITION_TYPES: &'static [PartitionType] =
   &[PartitionType::PARTITION_NONE, PartitionType::PARTITION_HORZ,
     PartitionType::PARTITION_VERT, PartitionType::PARTITION_SPLIT];
-
-pub static RAV1E_TX_TYPES: &'static [TxType] = &[
-  TxType::DCT_DCT,
-  TxType::ADST_DCT,
-  TxType::DCT_ADST,
-  TxType::ADST_ADST,
-  TxType::IDTX,
-  TxType::V_DCT,
-  TxType::H_DCT
-];
 
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub enum GlobalMVMode {
@@ -1119,30 +928,6 @@ impl PredictionMode {
       );
     }
   }
-}
-
-#[derive(Copy, Clone, PartialEq, PartialOrd)]
-pub enum TxSet {
-  // DCT only
-  TX_SET_DCTONLY,
-  // DCT + Identity only
-  TX_SET_DCT_IDTX,
-  // Discrete Trig transforms w/o flip (4) + Identity (1)
-  TX_SET_DTT4_IDTX,
-  // Discrete Trig transforms w/o flip (4) + Identity (1) + 1D Hor/vert DCT (2)
-  // for 16x16 only
-  TX_SET_DTT4_IDTX_1DDCT_16X16,
-  // Discrete Trig transforms w/o flip (4) + Identity (1) + 1D Hor/vert DCT (2)
-  TX_SET_DTT4_IDTX_1DDCT,
-  // Discrete Trig transforms w/ flip (9) + Identity (1)
-  TX_SET_DTT9_IDTX,
-  // Discrete Trig transforms w/ flip (9) + Identity (1) + 1D Hor/Ver DCT (2)
-  TX_SET_DTT9_IDTX_1DDCT,
-  // Discrete Trig transforms w/ flip (9) + Identity (1) + 1D Hor/Ver (6)
-  // for 16x16 only
-  TX_SET_ALL16_16X16,
-  // Discrete Trig transforms w/ flip (9) + Identity (1) + 1D Hor/Ver (6)
-  TX_SET_ALL16
 }
 
 pub fn has_tr(bo: BlockOffset, bsize: BlockSize) -> bool {
