+++ src/rdo_tables.rs
@@ -1,4 +1,3 @@
-
 pub const RDO_NUM_BINS: usize =  50;
 #[allow(unused)]
 pub const RDO_MAX_BIN: usize = 10000;
@@ -9,7 +8,7 @@ pub const RDO_QUANT_DIV: usize = 256/RDO_QUANT_BINS;
 pub const RDO_BIN_SIZE: u64 = (RDO_MAX_BIN / RDO_NUM_BINS) as u64;
 pub const RATE_EST_BIN_SIZE: u64 = (RATE_EST_MAX_BIN / RDO_NUM_BINS) as u64;
 
-use crate::partition::*;
+use crate::transform::TxSize;
 
 pub static RDO_RATE_TABLE: [[[u64; RDO_NUM_BINS]; TxSize::TX_SIZES_ALL]; RDO_QUANT_BINS] = [
 [[99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,99999,104,],
