+++ src/scan_order.rs
@@ -14,7 +14,7 @@
 
 const MAX_NEIGHBORS: usize = 2;
 
-use crate::partition::*;
+use crate::transform::*;
 
 #[repr(C)]
 pub struct SCAN_ORDER {
